Żeby w ogóle repo zadziałało musiałem zrobić update Android Studio do 3.0.1 i pobrać SDK 26 (nie wiem po co skoro pracujemy na niższym, ale IDE kazało). Potem konieczny był refresh gradle'a i ruszyło.
Przerzuciłem dotychczasowe MainActivity do BoardActivity. Teraz MainActivity to logowanie. Dodatkowo podpiąłem Trellolizer na onClick.
W sumie są 3 różne klucze i nie wiem który jest od czego:
1. MainActivity.secretApiKey - najdłuższy z kluczy, który dostałem od Trello
2. MainActivity.apiKey - drugi klucz, który dostałem od Trello
3. SharedPreferences.userAccessKey - klucz wklepany do formularza przy logowaniu
