+++ app/src/main/java/rustrll/sl/com/russiantrello/MainActivity.java
@@ -1,5 +1,9 @@
 package rustrll.sl.com.russiantrello;
 
+import android.app.Activity;
+import android.content.Intent;
+import android.content.SharedPreferences;
+import android.net.Uri;
 import android.support.design.widget.TabLayout;
 import android.support.design.widget.FloatingActionButton;
 import android.support.design.widget.Snackbar;
@@ -17,8 +21,10 @@ import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 
+import android.widget.EditText;
 import android.widget.ListView;
 
+import com.julienvey.trello.domain.Board;
 import com.julienvey.trello.domain.Card;
 
 import java.util.ArrayList;
@@ -26,229 +32,31 @@ import java.util.List;
 
 public class MainActivity extends AppCompatActivity {
 
-    /**
-     * The {@link android.support.v4.view.PagerAdapter} that will provide
-     * fragments for each of the sections. We use a
-     * {@link FragmentPagerAdapter} derivative, which will keep every
-     * loaded fragment in memory. If this becomes too memory intensive, it
-     * may be best to switch to a
-     * {@link android.support.v4.app.FragmentStatePagerAdapter}.
-     */
-    private SectionsPagerAdapter mSectionsPagerAdapter;
-    Trellolizer trellolizer;
-
-    /**
-     * The {@link ViewPager} that will host the section contents.
-     */
-    private ViewPager mViewPager;
+    static final public String secretApiKey = "8373585fdd1635ff23cd99aad95e450065c9c7d2d6a7deb691c55af87db5e622";
+    static final public String apiKey = "d18619238242c717df349f1a34dbead3";
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-
-
-        trellolizer = new Trellolizer();
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-        // Create the adapter that will return a fragment for each of the three
-        // primary sections of the activity.
-        mSectionsPagerAdapter = new SectionsPagerAdapter(getSupportFragmentManager());
-
-        // Set up the ViewPager with the sections adapter.
-        mViewPager = (ViewPager) findViewById(R.id.container);
-        mViewPager.setAdapter(mSectionsPagerAdapter);
-
-        TabLayout tabLayout = (TabLayout) findViewById(R.id.tabs);
-        tabLayout.setupWithViewPager(mViewPager);
-
-        FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-
-                String text = "";
-                List<Card> cards = trellolizer.get_cards();
-                for(int i = 0; i < cards.size(); i++) {
-                    text += cards.get(i).getName() + ", ";
-                }
-
-                Snackbar.make(view, text, Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-            }
-        });
-
+        EditText etToSave = (EditText) findViewById(R.id.editText1);
+        String textFromPreferences = getSharedPreferences("accessKey", Activity.MODE_PRIVATE).getString("editText1", "");
+        etToSave.setText(textFromPreferences);
     }
 
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_main2, menu);
-        return true;
+    protected void trelloRedirect(View v){
+        Uri uri = Uri.parse("https://trello.com/1/authorize?expiration=never&name=RussianTrello&key="+MainActivity.apiKey); // missing 'http://' will cause crashed
+        Intent intent = new Intent(Intent.ACTION_VIEW, uri);
+        startActivity(intent);
     }
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-
-        //noinspection SimplifiableIfStatement
-        if (id == R.id.action_settings) {
-            return true;
-        }
-
-        return super.onOptionsItemSelected(item);
-    }
-
-    /**
-     * A placeholder fragment containing a simple view.
-     */
-    public static class PlaceholderFragment extends Fragment {
-        /**
-         * The fragment argument representing the section number for this
-         * fragment.
-         */
-        private static final String ARG_SECTION_NUMBER = "section_number";
-
-        public PlaceholderFragment() {
-        }
-
-        /**
-         * Returns a new instance of this fragment for the given section
-         * number.
-         */
-        public static PlaceholderFragment newInstance(int sectionNumber) {
-            PlaceholderFragment fragment = new PlaceholderFragment();
-            Bundle args = new Bundle();
-            args.putInt(ARG_SECTION_NUMBER, sectionNumber);
-            fragment.setArguments(args);
-            return fragment;
-        }
-
-        @Override
-        public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                                 Bundle savedInstanceState) {
-            View rootView = inflater.inflate(R.layout.fragment_list, container, false);
-
-            ListView tasksListView = rootView.findViewById(R.id.tasksListView);
-
-            TaskAdapter tasksListAdapter = new TaskAdapter(getActivity().getApplicationContext(), tasksList(getArguments().getInt(ARG_SECTION_NUMBER)));
-            tasksListView.setAdapter(tasksListAdapter);
-            return rootView;
-        }
-
-
-        public ArrayList<Task> tasksList(int position) {
-            switch (position) {
-                case 5:
-                    return otherTasksList();
-                case 4:
-                    return mustTasksList();
-                case 3:
-                    return shouldTasksList();
-                case 2:
-                    return couldTasksList();
-                case 1:
-                    return wontTasksList();
-            }
-            return new ArrayList<>();
-        }
-
-        private ArrayList<Task> wontTasksList() {
-            ArrayList tasksList = new ArrayList<Task>();
-
-            tasksList.add(new Task("Zadanie 1", "Do jutra 12:00"));
-            tasksList.add(new Task("Zadanie 2", "Do dziś 13:00"));
-            tasksList.add(new Task("Zadanie 3", "Do jutra 9:00"));
-
-            return tasksList;
-        }
-
-        private ArrayList<Task> couldTasksList() {
-            ArrayList tasksList = new ArrayList<Task>();
-
-            tasksList.add(new Task("Zadanie 4", "Do jutra 12:00"));
-            tasksList.add(new Task("Zadanie 5", "Do dziś 13:00"));
-            tasksList.add(new Task("Zadanie 6", "Do jutra 9:00"));
-
-            return tasksList;
-
-        }
-
-        private ArrayList<Task> shouldTasksList() {
-            ArrayList tasksList = new ArrayList<Task>();
-
-            tasksList.add(new Task("Zadanie 7", "Do jutra 12:00"));
-            tasksList.add(new Task("Zadanie 8", "Do dziś 13:00"));
-            tasksList.add(new Task("Zadanie 9", "Do jutra 9:00"));
-
-            return tasksList;
-        }
-
-        private ArrayList<Task> mustTasksList() {
-            ArrayList tasksList = new ArrayList<Task>();
-
-            tasksList.add(new Task("Zadanie 10", "Do jutra 12:00"));
-            tasksList.add(new Task("Zadanie 11", "Do dziś 13:00"));
-            tasksList.add(new Task("Zadanie 12", "Do jutra 9:00"));
-
-            return tasksList;
-        }
-
-        private ArrayList<Task> otherTasksList() {
-            ArrayList tasksList = new ArrayList<Task>();
-
-            tasksList.add(new Task("Zadanie 13", "Do jutra 12:00"));
-            tasksList.add(new Task("Zadanie 14", "Do dziś 13:00"));
-            tasksList.add(new Task("Zadanie 15", "Do jutra 9:00"));
-
-            return tasksList;
-        }
-    }
-
-    /**
-     * A {@link FragmentPagerAdapter} that returns a fragment corresponding to
-     * one of the sections/tabs/pages.
-     */
-    public class SectionsPagerAdapter extends FragmentPagerAdapter {
-
-        public SectionsPagerAdapter(FragmentManager fm) {
-            super(fm);
-        }
-
-        @Override
-        public Fragment getItem(int position) {
-            // getItem is called to instantiate the fragment for the given page.
-            // Return a PlaceholderFragment (defined as a static inner class below).
-            return PlaceholderFragment.newInstance(position + 1);
-        }
-
-        @Override
-        public int getCount() {
-            // Show 3 total pages.
-            return 5;
-        }
-
-        @Override
-        public CharSequence getPageTitle(int position) {
-            switch (position) {
-                case 0:
-                    return "To Assign";
-                case 1:
-                    return "Must";
-                case 2:
-                    return "Should";
-                case 3:
-                    return "Could";
-                case 4:
-                    return "Won't";
-            }
-            return null;
-        }
+    protected void saveAccessKey(View v){
+        EditText etToSave = (EditText) findViewById(R.id.editText1);
+        SharedPreferences.Editor preferencesEditor = getSharedPreferences("userAccessKey", Activity.MODE_PRIVATE).edit();
+        String editTextData = etToSave.getText().toString();
+        preferencesEditor.putString("editText1", editTextData);
+        preferencesEditor.commit();
+        Intent boardIntent = new Intent(this, BoardActivity.class);
+        this.startActivity(boardIntent);
     }
 }
